Split of Tab components stories & renderer update
